requirements.txt
django
pytest
pytest-django
djangorestframework
django-cors-headers
python-decouple
pyyaml
uritemplate
